Fix deployment

+++ starter/main.py
@@ -18,18 +18,18 @@ if "DYNO" in os.environ and os.path.isdir(".dvc"):
     i = 0
     dvc_output = subprocess.run(
         ["dvc", "pull"], capture_output=True, text=True)
-    print(dvc_output.stderr)
-    print(dvc_output.stdout)
+    print(f'Std errors: {dvc_output.stderr}')
+    print(f'ST output: {dvc_output.stdout}')
     while dvc_output.stderr != 0 and i < 20:
         dvc_output = subprocess.run(
             ["dvc", "pull"], capture_output=True, text=True)
-        print(dvc_output.stderr)
-        print(dvc_output.stdout)
-
-        exit("dvc pull failed")
+        print(f'Std errors: {dvc_output.stderr}')
+        print(f'ST output: {dvc_output.stdout}')
         os.system("Retrying in 5 secs")
         i += 1
         time.sleep(5)
+    if i == 20:
+        exit("dvc pull failed")
 
     os.system("rm -r .dvc .apt/usr/lib/dvc")
